agent-ui/Agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "ServiceFlow AI Agent-UI Backend System"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv",
    "agno",
    "uvicorn[standard]",
    "fastapi",
    "schedule",
    "google-api-python-client>=2.176.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "beautifulsoup4",
    "spider-client",
    "tweepy",
    "requests",
    "charset-normalizer",
    "pymongo>=4.13.2",
    "nest-asyncio>=1.6.0",
    "pathlib>=1.0.1",
    "chardet>=5.2.0",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "discord.py>=2.3.0",
    "aiohttp",
    "langfuse",
    "openai",
    "cryptography>=42.0.0",
    "cffi>=1.16.0",
    "tzlocal>=5.0.0",
    "sqlalchemy>=2.0.0",
    "pandas>=2.0.0",
    "pypdf>=4.0.0",
    "openinference-instrumentation-agno",
    "fastmcp>=2.10.0",
]